requirements.txt
numpy
scipy
matplotlib
pygame-ce